Show in explorer won't work for "stake" txs
Resolves #352
