+++ src/lib/transactions/index.js
@@ -1,1 +1,2 @@
+export { default as getLastTransactionHash } from "./getLastTransactionHash";
 export { default as sortByHeightDesc } from "./sortByHeightDesc";
