+++ src/routes/(app)/dashboard/Contract/Operations/Send.svelte
@@ -1,12 +1,6 @@
 <script>
 	import { fade } from "svelte/transition";
 	import { mdiArrowUpBoldBoxOutline, mdiWalletOutline } from "@mdi/js";
-	import {
-		always,
-		compose,
-		getKey,
-		head
-	} from "lamb";
 
 	import { logo } from "$lib/dusk/icons";
 	import {
@@ -21,17 +15,11 @@
 	} from "$lib/dusk/components";
 	import { createCurrencyFormatter } from "$lib/dusk/currency";
 	import { operationsStore, settingsStore, walletStore } from "$lib/stores";
-	import { sortByHeightDesc } from "$lib/transactions";
+	import { getLastTransactionHash } from "$lib/transactions";
 
 	import GasSettings from "./GasSettings/GasSettings.svelte";
 	import OperationResult from "./OperationResult/OperationResult.svelte";
 	import ScanQr from "./ScanQR/ScanQR.svelte";
-
-	const getFirstHash = compose(getKey("id"), head);
-	const getLastTransactionHash = () => walletStore.getTransactionsHistory()
-		.then(sortByHeightDesc)
-		.then(getFirstHash)
-		.catch(always(""));
 
 	/** @type {Status[]} */
 	export let statuses;
@@ -211,16 +199,16 @@
 		</WizardStep>
 		<WizardStep step={3} {key} showNavigation={false}>
 			<OperationResult
+				errorMessage="Transaction failed"
 				onBeforeLeave={resetOperationStore}
 				operation={walletStore.transfer(address, amount).then(getLastTransactionHash)}
-				errorMessage="Transaction failed"
 				pendingMessage="Processing transaction"
 				successMessage="Transaction completed"
 			>
 				<svelte:fragment slot="success-content" let:result={hash}>
 					{#if hash}
 						<AnchorButton
-							href="https://explorer.dusk.network/transactions/transaction?id={hash}"
+							href={`https://explorer.dusk.network/transactions/transaction?id=${hash}`}
 							on:click={resetOperationStore}
 							rel="noopener noreferrer"
 							target="_blank"
